Collects a number of performance improvements.

Tests are done for a vector field with 200 x 200 x 10 cells.

# cutting plane
`plane('x')`
- old: 125 ms
- new: 248 μs

`plane('z')`
- old: 2.55s
- new: 1.26ms

# plotting

`plane('z').scalar()`
- old: 5.4 s  (no timeit)
- new: 2.85 s  (no timeit)
- new + new plane implementation: 34 ms

`plane('z').scalar(filter_field=field.norm)`
- old: 7.16s  (no timeit)
- new: 5.95s  (no timeit)
- new + new plane implementation: 47 ms